--- lib/server.js
@@ -1,66 +1,0 @@
-'use strict';
-
-const express = require('express');
-const app = express();
-const router = require('./cdRoutes');
-
-const PORT = process.env.PORT || 3000;
-
-let db = [];
-
-app.use(express.json());
-
-app.use( (req,res,next) => {
-  console.log('LOG:', req.method, req.path);
-  next();
-});
-
-app.get('/categories', (req,res,next) => {
-  let count = db.length;
-  let results = db;
-  res.json({count,results});
-});
-
-app.get('/categories/:id', (req,res,next) => {
-  let id = req.params.id;
-  let record = db.filter((record) => record.id === parseInt(id));
-  res.json(record[0]);
-  console.log('here',req.body);
-});
-
-app.post('/categories', (req,res,next) => {
-  let {name} = req.body;
-  let record = {name};
-  record.id = db.length + 1;
-  db.push(record);
-  res.json(record);
-});
-
-// Needs work I think
-app.put('/categories/:id', (req,res,next) => {
-  let id = req.params.id;
-  db.splice(id, 1);
-
-});
-
-app.delete('/categories/:id', (req,res,next) => {
-  let id = req.params.id;
-  console.log(db);
-  db.splice(id, 1);
-  console.log(db);
-  res.json(db);
-});
-
-app.use(router);
-
-app.get('*', (req,res) =>{
-  res.json('Not a valid route');
-});
-
-module.exports = {
-  server: app,
-  start: port => {
-    let PORT = port || process.env.PORT || 8080;
-    app.listen(PORT, () => console.log(`Listening on ${PORT}`));
-  },
-};
